chore: rename createProcessor to create-processor to match file convention
